Add workaround for global singleton; add specs (see more for info)
These specs cannot be run with browserify. Browserify will bundle all specs together as one file (which is nice and all but not needed for specs -- and actually limits singleton usages to one per karma configuration instead of test file). In order to run the singleton specs, comment the file back in. Focusing it should work. Also, the karma config could be modified to only include the singleton spec file. This is only a transitionary step while Mavenlink source code is converted completely to JavaScript modules and have zero dependencies on global variables.