Revert cache invalidation logic
Backgrid integrations in breaking because they assume the cache always exists. This is a bad integration since the UI should not care about the cache whatsoever. Alas, we have to revert to preserve existing features. If we want to improve the cache invalidation logic, we will want to remove any access to the cache first (which will identify these issues). And then bust the cache whenever we feel like it.