Make API endpoint host configurable
Allow an optional host to be passed into the StorageManager for when the
API endpoint to be used is on a different host. There is also an option
to set `withCredentials` to true for the XHR when using a different host.

This should be backwards compatible with existing usages of Brainstem.

Signed-off-by: Ellyx Jolley <ejolley@mavenlink.com>