Truly make the storage manager a singleton
By that I mean create the singleton storage manager once when the JS is initially evaluated, instead of providing a "get" method which will instantiate a new one if needed.

A "get" method is still implemented, though, so that we do not have to go update many many places calling "StorageManager.get".